requirements-install.txt
uvicorn
fastapi
sqlalchemy
alembic
psycopg2-binary
python-dotenv
cryptography
httpx
pydantic-settings
redis
python-multipart
